feat(asset-access): アセットアクセスのベースを作成
