+++ wasm/asset-access/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "asset-access"
+version = "0.1.0"
+edition.workspace = true
+authors.workspace = true
+repository.workspace = true
+rust-version.workspace = true
+
+[lib]
+crate-type = ["cdylib", "rlib"]
+
+[dependencies]
+wasm-bindgen.workspace = true
+wasm-utils = { workspace = true }
+webgl2 = { workspace = true, features = ["context", "shader"] }
+
+[dependencies.web-sys]
+workspace = true
+
+[profile.release]
+opt-level = "s"
