+++ wasm/asset-access/Makefile
@@ -1,0 +1,9 @@
+include ../../common.mk
+
+.PHONY: test
+test:
+	wasm-pack test --firefox --headless
+
+.PHONY: build
+build:
+	wasm-pack build -d ${ASSETS_DIR}/asset-access/pkg --target web
