+++ wasm/asset-access/src/lib.rs
@@ -1,0 +1,1 @@
+mod entry_point;
