+++ wasm/webgl2/src/lib.rs
@@ -21,6 +21,9 @@ pub mod font_asset;
 #[cfg(feature = "viewport")]
 pub mod viewport;
 
+#[cfg(feature = "shader")]
+pub mod shader;
+
 use error::Result;
 
 pub fn uniform_location(gl: &gl, program: &Program, name: &str) -> Result<WebGlUniformLocation> {
