+++ wasm/webgl2/src/shader/mod.rs
@@ -1,0 +1,1 @@
+pub mod texture;
