+++ web-server/assets/asset-access/index.html
@@ -1,0 +1,32 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <meta charset="utf-8">
+  <title>Blending</title>
+  <style>
+    body {
+      position: absolute;
+      top: 0;
+      left: 0;
+      width: 100%;
+      height: 100%;
+      display: flex;
+      flex-direction: column;
+      align-items: center;
+      justify-content: center;
+      padding: 0;
+      margin: 0;
+    }
+
+  </style>
+  <script type="module" src="./index.js"></script>
+  </head>
+
+<body>
+  <noscript>This page contains webassembly and javascript content, please enable javascript in your browser.</noscript>
+  <h2>WebGL Blending</h2>
+  <canvas id="webgl-canvas" style="background-color: black;"></canvas></canvas>
+</body>
+
+</html>
